+++ src/main/java/manyWorker/repository/SolicitudRepository.java
@@ -3,6 +3,6 @@ package manyWorker.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import manyWorker.entity.Solicitud;
 
-public interface SolicitudRepository extends JpaRepository<Solicitud, Long> {
+public interface SolicitudRepository extends JpaRepository<Solicitud, Integer> {
 	
 }
